requirements.txt
fastapi
uvicorn
joblib
Pillow
tensorflow
python-dotenv
firebase-admin
